AUT-1429: Allow sign in canary to run against integration
This allows for the sandpit canary to use the integration credentials for testing the sign in journey. This can make testing changes in sandpit easier, as there are fewer things to set up compared to the build environment, which sandpit tests normally run against.

Credentials need to be input into the canary-sign-in-with-ipv.tf file for testing locally, but not committed.

Co-authored-by: Bill Dunn <bill.dunn@digital.cabinet-office.gov.uk>