AUT-1429: Change sandpit cron offset
This changes the cron expression to run every 3 minutes but with an offset of 1 minute (this should mean it runs at 1, 4, 7, 10 etc minutes past the hour). This reduces the likelihood of it clashing with the integration canaries when we want to run the sandpit canaries against the integration backend. When running both integration and sandpit canaries against the integration backend at the same time, we can lead to e.g. otp codes from one run overwriting otp codes from another, leading to test failures.

Co-authored-by: Bill Dunn <bill.dunn@digital.cabinet-office.gov.uk>