AUT-1506: Turn on create account canaries in integration and production
Removing the start canary variable leaves it to the default, which is true, so this will start the create account journey. (I considered changing the variable to true but removing it is consistent with the other canaries)

These had been disabled since the switch from concouse to aws code pipelines, since a variable had not been migrated over, and the canaries were noisily failing. This has now been fixed, and we've trialled manually turning these alarms on in both environments, with successful runs.

The alarms will be turned on separately once we've seen a day or so of successful runs.