BAU: Fix Smoke test deployment
The previous change caused the workflow to fail and not deploy changes
to the smoke tests.

Instead of the previous method of pulling in the signed smoke test zips
during the `deploy` phase, instead, retrieve them immediately after the
signing job completes in `build`, and push them up to the GHA artifact
store. Then retrieve from this store in `deploy`.

This has two benefits:

1. The artifacts are easily downloadable from github, so developers
   can see exactly what it is that this workflow deployed
2. We no longer have to use the `aws s3 cp` globbing to retrieve the
   correct artifact version: we can just retrieve the specific key
   that the signing job returns.