BAU: Upgrade ESLint system for v9
A new major version of ESLint changes how it should be configured. A new `eslint.config.js` file replaces `.eslintrc.json` with tweaks to how configuration is done.

`eslint-plugin-n` is also brought in as a recommended fork of `eslint-plugin-node` that is no longer maintained. The new plugin adds a new rule that warns against use of `fetch` as it's not supported until Node 21. As it's already in place an exception is added to ignore the rule for `fetch`.

Further, unused disable directives in `alert.js` and `slack.js` are removed.