AUT-4043: Update smoke test numbers for sandpit canaries
* Makes sure that each of these numbers are different, meaning we can run the tests at the same time without risk of interference
* Uses the notify test numbers[1] to ensure that these do not trigger actual failed attempts to send to a number

This comes off the back of an incident where our smoke tests were using an offcom reserved number (07700900333) that is not also a notify test number. This was causing notify to attempt to send to this number but failing each time a smoke test using this number ran. At some point, we had enough failures to this number to trigger a warning from Notify, so we are fixing usages of this number to instead be one of the reserved notify test numbers.

This was largely a problem with the integration and production values which are stored in aws rather than the sandpit values which are changed here. The sandpit values are less of a problem since the sandpit canaries don't run by default, but updating here to save any future warnings.

[1] https://docs.notifications.service.gov.uk/rest-api.html\#phone-numbers